Edit clusterintalls reconcile to properly set cd conditions
The updated controller-runtime version requirement to explicitly set
status sub-resource uncovered an issue in
reconcileExistingInstallingClusterInstall in which the cd conditions
were not being updated.

Create a DeepCopy of the `conditions` variable, which was previously just
a pointer to cd.Status.Conditions. Pass this variable (instead of
cd.Status.Conditions) to controllerutils.SetClusterDeploymentCondition
so as not to mutate cd.Status.Conditions, but rather to append the new
conditions to the `conditions` variable.

Signed-off-by: Leah Leshchinsky <lleshchi@redhat.com>